requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
antlr4-python3-runtime
anyio
asttokens
attrs
backcall
backoff
beautifulsoup4
bleach
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
coloredlogs
comm
contourpy
cryptography
cycler
dataclasses-json
debugpy
decorator
defusedxml
Deprecated
dill
distro
docling
docling-core
docling-ibm-models
docling-parse
docopt
easyocr
effdet
elevenlabs
emoji
et_xmlfile
eval_type_backport
exceptiongroup
executing
faiss-cpu
fastapi
fastembed
fastjsonschema
ffmpeg-python
ffmpy
filelock
filetype
Flask
flatbuffers
fonttools
frozenlist
fsspec
future
gdown
gitdb
GitPython
google-api-core
google-auth
google-cloud-vision
google_search_results
googleapis-common-protos
greenlet
groovy
grpcio
grpcio-status
grpcio-tools
gunicorn
h11
h2
hpack
html5lib
httpcore
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
idna
imageio
importlib_metadata
importlib_resources
iniconfig
ipykernel
ipython
ipython_pygments_lexers
itsdangerous
jedi
Jinja2
jiter
joblib
jsonlines
jsonpatch
jsonpointer
jsonref
jsonschema
jsonschema-specifications
jupyterlab_pygments
kiwisolver
langchain
langchain-community
langchain-core
langchain-openai
langchain-qdrant
langchain-text-splitters
langdetect
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
latex2mathml
lazy_loader
loguru
lxml
Markdown
markdown-it-py
marko
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mistune
mmh3
mpire
mpmath
msgpack
multidict
multiprocess
mypy-extensions
narwhals
nbclient
nbconvert
nbformat
nest_asyncio
networkx
nibabel
nilearn
ninja
nltk
numpy
olefile
omegaconf
onnx
onnxruntime
openai
opencv-python
opencv-python-headless
openpyxl
orjson
packaging
pandas
pandocfilters
parso
pdfminer.six
pi_heif
pickleshare
pikepdf
pillow
pip
pipreqs
platformdirs
pluggy
portalocker
prompt_toolkit
propcache
proto-plus
protobuf
psutil
pure_eval
py_rust_stemmers
pyarrow
pyasn1
pyasn1_modules
pyclipper
pycocotools
pycparser
pydantic
pydantic_core
pydantic-settings
pydeck
pydub
Pygments
pylatexenc
pyparsing
pypdf
PyPDF2
pypdfium2
PySocks
pytest
python-bidi
python-dateutil
python-docx
python-dotenv
python-iso639
python-multipart
python-oxmsg
python-pptx
pytz
PyYAML
pyzmq
qdrant-client
RapidFuzz
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
rtree
ruff
safehttpx
safetensors
scikit-image
scikit-learn
scipy
seaborn
semantic-version
semchunk
sentence-transformers
setuptools
shapely
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
stack_data
starlette
sympy
tabulate
tenacity
threadpoolctl
tifffile
tiktoken
timm
tinycss2
tokenizers
toml
tomlkit
torch
torchvision
tornado
tqdm
traitlets
transformers
typer
typing_extensions
typing-inspect
typing-inspection
tzdata
urllib3
uvicorn
watchdog
wcwidth
webencodings
websockets
Werkzeug
wheel
wrapt
xlrd
XlsxWriter
yarg
yarl
zipp
zstandard